Change restrictions to added users for invitations
Here [0] permissions issue for stickers and media was fixed for newly
added users. But it was not actually an issue as chat itself even with
allowed media messages would block it.
But during investigation of this issue was found that permissions to
add users and send polls that usual users had was limited.

This patch fixes an issue described above.

[0] - 1e731f1a8e538062f12fb7e6d97b094a3ac6e9ba